Reduce the size of parameterized by two lines
parameterized did a type assetion to see if req.Body was a *parameters,
if so, it returned the value, if not it returned nil.

Howver the two arg form of a type asserting will assign to the first
parameter a zero value of the type asserted if the type does not
match. We can use this property to avoid the branch and make the
function two lines shorter.